pyproject.toml
[project]
name = "mylangchain1-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-ollama>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.3.3",
    "pypdf>=6.7.1",
]
